Debugueado todo excepto el producto de las matrices. Si estas leyendo esto, prometete a vos mismo una cosa. NUNCA... en TU VIDA, programes algo largo sin debuguearlo de a pedacitos. Ah y ademas,  USA UN IDE, DEBUGUEAR DESDE LINEA DE COMANDO ES LO PEOR QUESE INVENTO EN LA VIDA.

+++ matrix.h
@@ -8,8 +8,8 @@ float * makematrix (int rows, int column);  //creara una matriz con los elemento
 void randomatrix (float * matrix,int minnum,int maxnum);        //se llamara si se quiere inicializar con numeros random y si no, flaco llenala vos
 float * sumatrix (float * matrix1, float * matrix2);     //suma dos matrices, devuelve puntero a null si no se pueden sumar porque no son del mismo tamaño o no hay memoria suficiente para dar respuesta
 float * restamatrix (float * matrix1, float * matrix2); //lo mismo que la de suma, pero con resta
-float getvalmax (float * matrix);
-float getminval (float * matrix);
+int getvalmax (float * matrix);
+int getminval (float * matrix);
 int getfilas (float * matrix);
 int getcolumnas (float * matrix);               //estas 4 funciones devuelven lo pedido sobre matrices que estan en el mismo formato de esta libreria
 float * productomatrix(float * matA, float * matB);     //el programa hara: A.B=matC matC es el resultado
